+++ README.md
@@ -21,5 +21,8 @@ Implement Authentication
 npm install react-router-dom --> to switch between react component
 
 
+Adding Bootstrap
+ npm install bootstrap
+
