+++ todo-app/package.json
@@ -6,6 +6,7 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "bootstrap": "^5.3.2",
     "lint": "^0.8.19",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
