+++ todo-app/src/Components/todo/ErrorComponent.jsx
@@ -1,0 +1,10 @@
+export default function ErrorComponent() {
+    return (
+        <div className="ErrorComponent">
+            <h1>Oops! Something went wrong.</h1>
+            <div>
+                Apologies. Please, reach out to our team at ABC-DEF-GHIJ!
+            </div>
+        </div>
+    )
+}
