+++ todo-app/src/Components/todo/FooterComponent.jsx
@@ -1,0 +1,9 @@
+export default function FooterComponent() {
+    return (
+        <footer className="footer">
+            <div className='container'>
+                Copyright reserved @sagarsunar
+            </div>
+        </footer>
+    )
+}
