+++ todo-app/src/Components/todo/LogoutComponent.jsx
@@ -1,0 +1,8 @@
+export default function LogoutComponent() {
+    return (
+        <div className="LogoutComponent">
+          <h1>You are logged out.</h1>
+           <div>Thank you for visiting us. See you soon!</div>
+        </div>
+    )
+}
