+++ todo-app/src/Components/todo/TodoApp.css
@@ -1,4 +1,11 @@
-label, input, button{
+label, input, button {
     margin: 10px;
     padding: 5px;
+}
+    
+footer {
+    bottom: 0;
+    position: absolute;
+    width: 90%;
+    height: 40px;
 }
